install python-netifaces

+++ deploy_filesystem.sh
@@ -14,9 +14,11 @@ cp -rvf $1/squashfs_override/* ./
 mount -t proc none ./proc
 mount devpts /dev/pts -t devpts
 cp /etc/resolv.conf ./etc/resolv.conf
-chroot . ash -c 'tazpkg get-install python bash clonezilla perl-uri linux-nfsd nfs-utils locale-fr'
+chroot . ash -c 'tazpkg get-install python python-netifaces bash clonezilla perl-uri linux-nfsd nfs-utils locale-fr'
 for file in {python-tftpy-0.8.0.tazpkg,fusioninventory-agent-2.4.2.tazpkg,perl-universal-require-0.18.tazpkg,perl-file-which-1.22.tazpkg,perl-treepp-0.43.tazpkg,python-psutil-5.4.3.tazpkg}; do
-    curl -O https://agents.siveo.net/imaging/${file}
+    if [[ ! -f $file ]]; then
+        curl -O https://agents.siveo.net/imaging/${file}
+    fi
 done
 chroot . ash -c 'tazpkg install *.tazpkg'
 chroot . ash -c 'adduser -D siveo'
